requirements.txt
requests>=2.28.0
textblob>=0.17.0
scikit-learn>=1.0.0
numpy>=1.21.0
pandas>=1.3.0
flask>=2.0.0
flask-cors>=4.0.0
python-telegram-bot>=20.0
schedule>=1.2.0
python-dotenv>=1.0.0


python-ta>=0.3.0

